data-ingestion/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
anyio==4.11.0
    # via starlette
blinker==1.9.0
    # via flask
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.11.0
    # via functions-framework
colorama==0.4.6
    # via click
deprecation==2.1.0
    # via cloudevents
flask==3.1.2
    # via functions-framework
functions-framework==3.9.2
    # via data-ingestion (pyproject.toml)
google-api-core==2.27.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.4.1
    # via data-ingestion (pyproject.toml)
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.71.0
    # via google-api-core
gunicorn==23.0.0
    # via uvicorn-worker
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via functions-framework
typing-extensions==4.15.0
    # via
    #   anyio
    #   starlette
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.3
    # via
    #   flask
    #   functions-framework
